+++ Gemfile
@@ -1,0 +1,5 @@
+source "http://rubygems.org"
+
+gemspec
+
+gem "rake"
