+++ shoden.gemspec
@@ -1,6 +1,6 @@
 Gem::Specification.new do |s|
   s.name              = "shoden"
-  s.version           = "0.4.1"
+  s.version           = "1.0"
   s.summary           = "Object hash mapper for postgres"
   s.description       = "Slim postgres models"
   s.authors           = ["elcuervo"]
@@ -10,6 +10,7 @@ Gem::Specification.new do |s|
   s.files             = `git ls-files`.split("\n")
   s.test_files        = `git ls-files test`.split("\n")
 
-  s.add_dependency("sequel", "~> 4.11.0")
-  s.add_development_dependency("cutest", "~> 1.2.1")
+  s.add_dependency("pg", "~> 1.1")
+
+  s.add_development_dependency("cutest", "~> 1.2")
 end
